Fix role creation, start preparing service tests
